moved function to helper file
